refactor: Add FeatureBoard for configuring Isthmus (#101)
* Add FeatureBoard for configuring Isthmus

This change adds a FeatureBoard, a collection of configuration and flags
to control the behavior of SQL to Substrait conversion and vice versa.
The FeatureBoard replaces current Options based approach and is inpired
by Feature Toggle design pattern. The new implementation also provides
additional configurability using default config files (see picocli
PropertiesDefaultProvider).

Removing picocli.CommandLine.PropertiesDefaultProvider as its
constructor is not found in the nativeImage. This fix should unblock the
build for now.