fix(spark): incorrect conversion of expand relation
In the expand relation, the projection expressions are stored in a
two dimensional array.  The spark matrix needs to be transposed
in order to map the expressions into substrait, and vice-versa.  I hadn’t noticed this earlier.

Also, the remap field should not be used because the outputs
are defined directly in the projections array.

Signed-off-by: Andrew Coleman <andrew_coleman@uk.ibm.com>