fix(isthmus): concat of different string types
The type matching used when mapping Calcite to Substrait expressions
required parameters and return type of the concat function to be either
all varchar or all string. Any mixing of types caused a failure.

This change relaxes the type matching to allow string types (varchar,
char and string literal) to be used interchangeably.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>