fix(spark): enable aliased expressions to round-trip
A number of the TPC-DS tests were failing because the query
contains multiple aliases to the same expression, causing a
potential mismatch in the reference index.  Although the plans
were equivalent, the substrait POJO comparison failed.
This commit uses the `hint` field of the Rel message to store the
alias names, and restore them back to the Spark plan to match
the original.