fix(spark): use ImmutableFileFormat builders
Round-trip testing was failing for query plans that involved
reading files from the filesystem.
This was due to the conversion code instantiating objects
directly rather than using the immutable builders.
The round-trip object comparison was then failing.
This commit fixes that by using the ImmutableFileFormat
builders, and hence the generated equals() methods.