build: use Gradle signing instead of JReleaser
JReleaser expects to have the corresponding public key available when
signing artifacts so that it can verify that the public key is published
to known key registries. While publishing works without the public key,
it outputs a warning message, which can cause confusion when viewing
build logs.

This change disabled JReleaser artifact signing and reverts to using
Gradle's own signing plugin. This also provides a smoother migration
path to an official Sonatype Gradle publishing plugin, if/when one
becomes available in the future.

Gradle wrapper is also updated to the latest patch release.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>