feat(core): uri -> urn + add urn validation
This commit fully moves from URIs to URNs. This is meant as an
intermediate commit in the graceful migration. Later work will re-add
support for URIs in a way such that both URIs and URNs are accepted on
parse and emitted in plans. It also adds necessary kotlin dep to make debugging work.

BREAKING CHANGE: this commit alters the API by dropping support for
URIs and adding support for URNs.