ci: handle warnings in compileScala as errors (#519)
Enable the `-Xfatal-warnings` flag for the Scala compiler to
turn warnings into errors (i.e. fail the build).
Suppress this warning for the case where the spark converter
is handling deprecated substrait enum values in the Join relation.
We still want the converter to hande this legacy value without
throwing an exception.

Signed-off-by: Andrew Coleman <andrew_coleman@uk.ibm.com>