fix(spark): sporadic failures in Hive test suite (#540)
The builds were sporadically failing with the error:
io.substrait.spark.HiveTableSuite > Insert into Hive table FAILED
    org.scalatest.exceptions.TestFailedException: Expected 4, but got 5

It turns out that the row values were not being correctly
serialised / deserialised into the Hive table correctly, so even in
the cases were the test passed (based on number of rows
added), the values inserted were still incorrect.
This commit fixes that, and adds an extra check in the test
case(s) to assert the correct values are inserted.

Signed-off-by: Andrew Coleman <andrew_coleman@uk.ibm.com>