build: replace jreleaser with gradle-maven-publish
There are two issues with JReleaser:

1. It is a heavy weight plugin designed to manage the entire release
   process, similar to semantic-release, and it is only used for
   publishing to the Maven Central Publisher API.
2. Transitive dependency conflicts between Spotless 8.x and JRelease 1.x
   plugins prevent them being used together.

This change replaces JReleaser with the vanniktech gradle-maven-publish
plugin. This is a lighter weight plugin that provides a replacement for
the standard maven-publish plugin, adding the capability to publish to
Maven Central.

This change also updates Spotless to 8.0.0 since conflicts with
JReleaser no longer occur.

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>