feat(spark): dialect YAML file for spark converter (#554)
Adds the formal dialect definition for the Spark converter.
The file is generated from code which extracts the relevant
information from the exisitng FunctionMapper definitions.
This ensures that the dialect YAML remains in-sync with the
code.
Tests have been introduced to ensure that the dialect conforms
to the schema (defined in teh base substrait library), and that
the published dialect YAML is the equivalent to one that would
be generated from the current code.

Signed-off-by: Andrew Coleman <andrew_coleman@uk.ibm.com>