fix: prevents exception on construction of `FunctionConverter` with duplicate functions (#564)
* test: introduce test to capture bug #562

This commit introduces a test to capture the bug where name-colliding
extension functions with different URNs cause an exception to be
thrown on `FunctionConverter` construction.

* refactor: rename alm to nameToFnMap in FunctionConverter

This is a small change to make the meaning of a variable clearer in
the `FunctionConverter.java` implementation.

* fix: allow multiple same-named functions in FunctionConverter

* test: that order added determines precedence in FunctionConverter

* test: simplify tests and add special test for overlapping default ext

Added a test for ltrim to ensure there is no issue specifically with
the default extension collection functions (which have some special
handling in isthmus) and an introduced ltrim function.

* fix: address @nielspardon comments on PR #564

This commit:
- improves the documentation around the `FunctionConverter.java` class
- fixes some minor stylistic comments around the tests

* docs: wrap mentions of SqlOperator in @link