chore(isthmus): remove dead code
Since the use of DdlSqlToRelConverter was added to SubstraitSqlToCalcite
in commit 260a1c42f7d1ffc771ee50b481972ca95ee3e4e4, the original
invocation of SqlToRelConverter that was left as a fallback when it
returns null is never called. The reason is that DdlSqlToRelConverter
itself already falls back to calling SqlToRelConverter so it never
returns null. This means that removeRedundantProjects is never applied
to the Rel structure since this only occurs in the uncalled fallback
path. All tests continue to pass, which indicates that this optimiation
is unnecessary.

If it required in the future, it might be better written as follows to
provide flexibility in the rules applied:

static RelRoot applyPlannerRules(RelRoot root, RelOptRule... rules) {
  HepProgramBuilder builder = HepProgram.builder();
  for (RelOptRule rule : rules) {
    builder.addRuleInstance(rule);
  }

  HepPlanner planner = new HepPlanner(builder.build());
  planner.setRoot(root.rel);

  RelNode optimizedExpression = planner.findBestExp();

  return root.withRel(optimizedExpression);
}

Signed-off-by: Mark S. Lewis <Mark.S.Lewis@outlook.com>