feat(isthmus): support month()-style datetime operators (#643)
The Substrait Function for Month style mappings doesn't work as the
calcite only has 2 arguments, but substrait has three.

Isthmus (in either direction) assumes that there is a 1:1 mapping betwen a
substrait function and a calicte function. With the same number of arguments
in each. (although types can be coerced if needeed)

In this case, the 3 arg Substrait function needs to be mapped to a 2 arg
calcite one. In effect this is working by swapping out the function to a different
one and then letting the rest of the logic work as before.

A concern is the increased complexity - and the code as a whole wasn't built with this in mind.

Signed-off-by: MBWhite <whitemat@uk.ibm.com>